refactor(templates): isolate translation template tags
Backend modules only need selected rendering helpers, but importing the template-tag library loads its full dependency graph. Move shared helpers to dedicated modules while preserving the template interface.
